requirements.txt
streamlit

langchain-community
langchain-groq

chromadb==0.3.21
python-dotenv==0.19.2
pandas

pyyaml>=6.0.2,<7.0.0
beautifulsoup4